Remove asssert in OLI test 1
We cannot use this assert, becasue the returned value of res is:

ERROR: Account test: no folder to sync (f[388 chars]\n\n' != ''
- ERROR: Account test: no folder to sync (folderfilter issue?)
- ERROR: Exceptions occurred during the run!
- ERROR: Account test: no folder to sync (folderfilter issue?)

So, if we have an assert to test if the string is empty,
the test stops here. We can check if we have or not folder
in the next  check (count_maildir_mails().