Remove asssert in OLI test 2
We cannot use this assert, becasue the returned value of res is:

ERROR: getfolder() asked for a nonexisting folder 'INBOX.OLItest'.
ERROR: Exceptions occurred during the run!
ERROR: getfolder() asked for a nonexisting folder 'INBOX.OLItest'.

So, if we have an assert, to compare if the string is empty,
the test stops here. We can check if we have or not folder in the
next  check (count_maildir_mails().