Handle socks connection errors
This patch handles the socks connection problems. First, move the "msg"
variable to the places where the variable is used. This variable could
be unset in the for loop, so is better move to these places.

This patch also removes the "msg" variable set in th except blocks, like
except socket.error as msg, because the variable is not used.

The patch also removes the "i" variable, not used.

Finally, the patch correct some typos.

Close: #67

Signed-off-by: Rodolfo García Peñas (kix) <kix@kix.es>