imap: always use cached capabilities
`imapobj` has a function call `capability()` and a cached version of
that `capabilities`. The call of function literally makes a command
`CAPABILITY` to IMAP server and some IMAP server may limits the number
of that command to one that leads to fail like:
```
 ERROR: While attempting to sync account 'Personal'
  imaplib2.imaplib2.IMAP4.abort: command: CAPABILITY => socket error: <class 'OSError'> - Too many read 0
```

Modern version of imaplib2 fetches it when connection is established.

Signed-off-by: Kirill A. Korinsky <kirill@korins.ky>