Handle exceptions thrown while getting message ID
Python's email library can throw both an IndexError as well as a
HeaderParseError when the email has certain invalid data in the
message ID. This prevents users with faulty emails from syncing.
This change allows us to simply continue without a message ID as
was intended in the code before anyway.

Signed-off-by: Florian Snow <florian@familysnow.net>