offlineimap.conf: use raw strings in regex examples
Newer pythons are strict on the escape sequences allowed in normal
strings, which can cause issues with regular expressions as most of them
do include escape sequences that are not allowed. In those cases, the
process will exit with the confusing warning:
  <string>:1: SyntaxWarning: invalid escape sequence '\.'

To help prevent such cases, have the sample configuration use raw
strings when using regular expressions.

Signed-off-by: Luciano Rocha <lfrocha@gmail.com>