avoid duplicate install of bin/offlineimap
fixes install with newer versions of tools, which otherwise fails with `FileExistsError: File already exists` in installer/destinations.py write_file (duplicate install in pyproject.toml [project scripts]). I think this is the "Finally, the
`bin/offlineimap` command is not used, so we probably can remove it" referred to in 4f5b252 (which wasn't done at the time but now seems to be required).